feature: Background colour change on windows
Moved around some code from DataViewer to make use of pre-made methods in Controller.java.

Changed the way in which map's heat map is calculated. Uses new deaths (same as before) but the base value is now the maximum of the new deaths to make the map relative and represent accurately.