fix: Changing Dates functioning as intended
date change method is defined in super class Controller and calls 'processDataInDateRange' which is specific for each controller.

eg. switching from main panel -> map panel
when in main panel, it would invoke the 'processDataInDateRange' as it should. Upon switching to map panel it carries the date from the previous panel over. However, if the date is changed in the map panel, no changes will be shown, because we are still in the DataViewerController class, and changing date would invoke its processDataInDateRange not map panel's.

Fix: Overriding the dateChanged method in the DataViewerController to call the 'processDataInDateRange' of the current controller instead.