Functionality for fadeTransition is complete for StatsWindow. This actually simplified the code for "forwardButton" and "backwardButton".
This works by setting the initial opacity of each statistic panel to 0 instead of setting visibility to false and then using fades to change what panel can be seen. "Visibility == false" does not allow fade transitions to work, so this was the best alternative.

"fadeOut.setNode(statsPanes.get((panelIndex) % statsPanes.size()));" cannot have "panelIndex - 1" without a check for a negative index like in the backwardButton